[solvers.hh] Always convert the reduction to double explicitly.

See merge request core/dune-istl!138

(cherry picked from commit e8223df82ab14e49c1a0dff07fc3bc6e75e98851)

82476058 [solvers.hh] Always convert the reduction to double explicitly.

+++ dune/istl/solvers.hh
@@ -131,7 +131,7 @@ namespace Dune {
 
       // fill statistics
       res.iterations = i;
-      res.reduction = max_value(def/def0);
+      res.reduction = static_cast<double>(max_value(def/def0));
       res.conv_rate  = pow(res.reduction,1.0/i);
       res.elapsed = watch.elapsed();
 
@@ -1516,7 +1516,7 @@ namespace Dune {
 
       // fill statistics
       res.iterations = i;
-      res.reduction = max_value(def/def0);
+      res.reduction = static_cast<double>(max_value(def/def0));
       res.conv_rate  = pow(res.reduction,1.0/i);
       res.elapsed = watch.elapsed();
